feat(constant): api path 및 아이콘 상수화
